move all salesgpt files to folder
