+++ app/salesGPT/components/CV.module.scss
