+++ app/salesGPT/components/CV.tsx
