+++ app/salesGPT/components/employeeCVSummary.module.scss
@@ -1,4 +1,4 @@
-@import "./window.scss";
+@import "../../components/window.scss";
 
 .input {
   display: flex;
