+++ app/salesGPT/components/employeeCard.module.scss
