+++ app/salesGPT/components/employeeCard.tsx
