+++ app/salesGPT/components/salesGPT.module.scss
@@ -1,4 +1,4 @@
-@import "./window.scss";
+@import "../../components/window.scss";
 
 @mixin container {
   // --sidebar-width: 50%;
